django-page factory and page object
this moves all the logic for creating an html doc with embedded JSON
output for ember into the ember-page factory. the page object is a thin
wrapper around that factory which handles bootstrapping the output

this allows us to interact with a django page to retrieve output as we'd
expect the server to produce prior to ember booting.