pass in requestType from manual serialization so we can call _super
channels are undergoing a slight change in how they serialize in order
to compensate for updates from the server to provide keys in dasherized
format instead fo camel case. this means we need to call super in the
channel serializer, which leads to a few other changes in how other
models are serialized.